E2E deploy test: rewrite in go tests (#814)
* :sparkles: add tests/e2e package


* :sparkles: use stream apis


* :sparkles: add GetObject & SearchByID


* :recycle: improve validation


* :green_heart: update e2e-deploy


* :green_heart: add libhdf5 for e2e-deploy test


* :green_heart: fix vald client


* :green_heart: fix NewPortforward interface


* :fire: commented out logging search results


* :white_check_mark: improve e2e test


* :bento: update values-ci manifest


* :green_heart: simplify


* :wrench: add config


* :wrench: inc replicas


* :green_heart: fix loop


* :loud_sound: add log


* :mag: improve logs


* :wrench: add keepalive


* :green_heart: add Cassandra E2E test


* :wrench: revise config


* :wrench: revise config


* :wrench: cassandra consistency = one


* :green_heart: trigger e2e deploy test using chatops


* :white_check_mark: use correct version of containers


* :loud_sound: print HELM_EXTRA_OPTIONS


* :green_heart: remove redundant condition checks


* :arrow_up: upgrade setup-k3d


* :white_check_mark: just a workaround for strange mysql table definition


* :truck: move values into .github/helm/values


* :white_check_mark: add sidecar test


* :wrench: fix


* :wrench: fix config for agent-sidecar


* :wrench: reduce number of data


* :wrench: add createindex phase


* :wrench: remove useless config


* :green_heart: remove useless lines


* :wrench: fix sidecar config


* :white_check_mark: add tests for mysql deleteMeta


* :wrench: Add options for which data is used for test


* :art: format yamls


* :page_facing_up: Update license header


* :arrow_up: Upgrade to v1 APIs


* :wrench: Fix Helm values for E2E test


* :green_heart: Fix method call


* :wrench: Reduce number of data


* :green_heart: Fix to use v1 vald API


* :wrench: Reduce number of Scylla cluster member


* :children_crossing: Add a variable for scylla deploy task


* :wrench: Add minio deploy task


* :recycle: Use kubectl wait instead for sh loop


:pencil2: Fix typo


* :heavy_plus_sign: Add cli-runtime to go.mod.default / Remove duplicated client-go entry


* :recycle: use assoc list for checking pr-xxx image existences


* :arrow_up: Update go.sum


* :green_heart: Fix build


* :wrench: tuning: add sleep


* :wrench: Tuning number of requests


* :green_heart: Fix e2e tests code


* :white_check_mark: Update e2e tests

+++ internal/db/rdb/mysql/mysql.go
@@ -349,7 +349,7 @@ func (m *mySQLClient) SetVectors(ctx context.Context, vecs ...Vector) error {
 	return tx.Commit()
 }
 
-func (m *mySQLClient) deleteVector(ctx context.Context, val interface{}) error {
+func (m *mySQLClient) deleteVector(ctx context.Context, val string) error {
 	if !m.connected.Load().(bool) {
 		return errors.ErrMySQLConnectionClosed
 	}
@@ -363,12 +363,21 @@ func (m *mySQLClient) deleteVector(ctx context.Context, val interface{}) error {
 	}
 	defer tx.RollbackUnlessCommitted()
 
+	var id int64
+	_, err = tx.Select(idColumnName).From(vectorTableName).Where(m.dbr.Eq(uuidColumnName, val)).Limit(1).LoadContext(ctx, &id)
+	if err != nil {
+		return err
+	}
+	if id == 0 {
+		return errors.ErrRequiredElementNotFoundByUUID(val)
+	}
+
 	_, err = tx.DeleteFrom(vectorTableName).Where(m.dbr.Eq(uuidColumnName, val)).ExecContext(ctx)
 	if err != nil {
 		return err
 	}
 
-	_, err = tx.DeleteFrom(podIPTableName).Where(m.dbr.Eq(uuidColumnName, val)).ExecContext(ctx)
+	_, err = tx.DeleteFrom(podIPTableName).Where(m.dbr.Eq(idColumnName, id)).ExecContext(ctx)
 	if err != nil {
 		return err
 	}
@@ -381,8 +390,15 @@ func (m *mySQLClient) DeleteVector(ctx context.Context, uuid string) error {
 }
 
 // DeleteVectors is the same as DeleteVector() but it deletes multiple records.
-func (m *mySQLClient) DeleteVectors(ctx context.Context, uuids ...string) error {
-	return m.deleteVector(ctx, uuids)
+func (m *mySQLClient) DeleteVectors(ctx context.Context, uuids ...string) (err error) {
+	for _, uuid := range uuids {
+		err = m.deleteVector(ctx, uuid)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
 }
 
 // SetIPs insert the vector's uuid and the podIPs into database.
